Make sure page changes and body classes are in sync.
On complex styled pages (or slow hardware), there could be a flicker,
because it was possible for the body to lose the `.transitioning` class 
in the period before the relevant `Meteor.flush` that would be triggered
by the transition end.

Here, I trigger it manually, but I'd prefer to just wait until flush-time
to change the body class. (OR better yet, just set the body class with a 
reactive var).